refactor(scripts): split cleanup tools
Move city onboarding evaluation and Laserfiche repair logic into focused modules. Preserve CLI behavior, stdout contracts, helper aliases, and patch seams used by existing operator tests.
